+++ src/algo/registration/embedded_deformation/ed.h
@@ -68,7 +68,7 @@ private:
 public:
 	bool finished();
 	bool solveIteration() override;
-	int  currentIteration() override;
+	size_t currentIteration() override;
 	bool solve() override;
 public:
 	const SurfaceMesh & getSource() override;
